+++ R/dplyr.R
@@ -61,12 +61,15 @@ tbl.BigQueryConnection <- function(src, from, ...) {
   }
 
   # This is ugly, but I don't see a better way of doing this
-  tb <- as_bq_table(src$con, from)
-  tbl$lazy_query$is_view <- !inherits(from, "sql") &&
-    tryCatch(
+  if (!inherits(from, "sql")) {
+    tb <- as_bq_table(src$con, from)
+    tbl$lazy_query$is_view <- tryCatch(
       bq_table_meta(tb, "type")$type == "VIEW",
       bigrquery_notFound = function(e) TRUE
     )
+  } else {
+    tbl$lazy_query$is_view <- FALSE
+  }
   tbl
 }