+++ R/bq-projects.R
@@ -1,12 +1,12 @@
 #' List available projects
 #'
 #' List all projects that you have access to. You can also work with
-#' [public datasets](https://cloud.google.com/bigquery/public-data/),
+#' [public datasets](https://docs.cloud.google.com/bigquery/public-data/),
 #' but you will need to provide a `billing` project whenever you perform
 #' any non-free operation.
 #'
 #' @section Google BigQuery API documentation:
-#' * [list](https://cloud.google.com/bigquery/docs/reference/rest/v2/projects/list)
+#' * [list](https://docs.cloud.google.com/bigquery/docs/reference/rest/v2/projects/list)
 #' @param page_size Number of items per page.
 #' @param max_pages Maximum number of pages to retrieve. Use `Inf` to retrieve
 #'   all pages (this may take a long time!)