+++ NEWS.md
@@ -3,6 +3,7 @@
 * `bq_perform_upload()` and friends now default to 22 digits of accuracy, and now allow you to change this value with the new `json_digits` argument.
 * Always upload `POSIXt` objects with 6 digits (i.e. microsecond) precision  (#660). 
 * Correctly set the `"tzone"` attribute of uploaded `POSIXt` columns to `Sys.timezone()` when the timezone is empty or not defined (@r2evans, #660). 
+* `dbAppendTable()` now correctly uses the billing project from the connection (@eheinzen, #638).
 * `dbExecute()` gains `params=` support (@r2evans, #667).
 * `dbExecute()` and `dbSendQuery()` error if you accidentally use `parameters`  (@r2evans, #667).
 * Check `getOption("bigrquery.quiet")` option in more `bq_*` functions (@r2evans, #663).