+++ NEWS.md
@@ -4,6 +4,7 @@
 * Always upload `POSIXt` objects with 6 digits (i.e. microsecond) precision  (#660). 
 * Correctly set the `"tzone"` attribute of uploaded `POSIXt` columns to `Sys.timezone()` when the timezone is empty or not defined (@r2evans, #660). 
 * `dbAppendTable()` now correctly uses the billing project from the connection (@eheinzen, #638).
+* `dbCreateTable()` and `dbWriteTable()` now correctly say that `temporary = TRUE` (rather than `FALSE`) is unsupported (#670).
 * `dbExecute()` gains `params=` support (@r2evans, #667).
 * `dbExecute()` and `dbSendQuery()` error if you accidentally use `parameters`  (@r2evans, #667).
 * Check `getOption("bigrquery.quiet")` option in more `bq_*` functions (@r2evans, #663).